Add password.reset command.
This command will interactively ask the user for their email address,
and then wait for them to enter the password reset token we sent to
their email.

Use `password.reset --continue` if you already have a password reset
token. This handles the case when the user quits the CLI before entering
the password reset token.